Extract RPG sheet logic into dedicated service
Move RPG character sheet rules, payload normalization, validation, and PDF rendering from `RpgCharEditorController` into a new `RpgCharacterSheetService`, and wire both RPG controllers plus the editor view to use it as the single source of truth. Update feature/unit tests to reference the service and add coverage for slot-limit failures so portrait files are only written after slot checks pass inside the transaction.
